+++ tests/ts/ads/dfpService.test.ts
@@ -1,4 +1,4 @@
-import { createDom } from '../stubs/browserEnvSetup';
+import { createDom, dom } from '../stubs/browserEnvSetup';
 import { expect, use } from 'chai';
 import * as sinonChai from 'sinon-chai';
 import * as Sinon from 'sinon';
@@ -2340,6 +2340,34 @@ describe('DfpService', () => {
         consent: consentConfig,
         logger: noopLogger
       }).then(() => {
+        expect(setNonPersonalizedAdsSpy).to.be.calledOnce;
+        expect(setNonPersonalizedAdsSpy).to.be.calledOnceWithExactly(0);
+      });
+    });
+
+    it('should use the cmp timeout for faktor.io', () => {
+      const timeout = 10;
+      dom.window.__cmp = (cmd: string, params: any, callback: Function) => {
+        console.log('calling', cmd, params);
+        return;
+      };
+      const cmpSpy = sandbox.spy(dom.window, '__cmp');
+
+      return newDfpService().initialize({
+        slots: [],
+        consent: {
+          ...consentConfig,
+          cmpConfig: {
+            provider: 'faktor',
+            autoOptIn: true,
+            timeout: timeout
+          }
+        },
+        logger: noopLogger,
+      }).then(() => {
+        expect(cmpSpy).to.be.calledOnce;
+        expect(cmpSpy.firstCall.args[0]).to.be.equal('addEventListener');
+        expect(cmpSpy.firstCall.args[1]).to.be.equal('cmpReady');
         expect(setNonPersonalizedAdsSpy).to.be.calledOnce;
         expect(setNonPersonalizedAdsSpy).to.be.calledOnceWithExactly(0);
       });